requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
faiss-cpu
pydantic
